+++ VirtualRouter/src/protocols/eigrp/rtp/EigrpPacketBuilder.h
@@ -24,7 +24,7 @@ class EigrpPacketBuilder
 {
 public:
     static std::optional<EigrpHeader> buildHeader(PacketBuilder& packet,
-        uint8_t opcode, uint32_t seq, uint32_t ack, uint16_t virId, uint16_t asn);
+        uint8_t opcode, uint32_t seq, uint32_t ack, uint16_t virId, uint32_t asn);
     static bool appendParameterTLV(TLV16BufferManager& tlv, EigrpInterface& iface);
     static bool appendVersionTLV(TLV16BufferManager& tlv);
     static bool appendMulticastSeqTLV(TLV16BufferManager& tlv, uint32_t seq);
